% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_totals}
\alias{add_totals}
\title{Columnas/Filas Totales}
\usage{
add_totals(dt, axis = "columns", by, keys)
}
\arguments{
\item{dt}{data.table a aplicar la funcion}

\item{axis}{Si los totales son por columna ('columns') o si son por tupla ('rows')}

\item{by}{Columnas que correpsonden a la agregacion.}

\item{keys}{Columnas que corresponden a llaves de identidad.}
}
\value{
data.table con columnas o tuplas con totales.
}
\description{
Genera totales para columnas o para tuplas de acuerdo a agregaciones.
}
