% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assure_type}
\alias{assure_type}
\title{Castea variables entre data.frames}
\usage{
assure_type(DT, var, var2, keep.y = TRUE)
}
\arguments{
\item{DT}{Data.table a realizar el casteo de variables}

\item{var}{Vector de caracteres correspondiente a variables a castear.}

\item{var2}{Vector de caracteres correspondiente a variables a castear.}

\item{keep.y}{Si se quiere conservar el tipo de var2 (TRUE) o de var (FALSE)}
}
\value{
DT con variables casteadas.
}
\description{
Se encarga de castear variables al mismo tipo. Si keep.y esta
activado, var se castea al tipo  de var2, de lo contrario var2 se castea al
tipo de var.
}
\note{
Solo castea a valores primitivos.
}
