% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assure_type_DTs}
\alias{assure_type_DTs}
\title{Asegura tipos entre data.tables}
\usage{
assure_type_DTs(xDT, yDT, keys, keep.y = T)
}
\arguments{
\item{xDT}{data.table en la izquierda}

\item{yDT}{data.table en la derecha}

\item{keys}{Vector de caracteres correspondiente a llaves a castear.}

\item{keep.y}{Si se da prioridad a yDT (TRUE) o a xDT(FALSE) en el casteo.}
}
\value{
data.tables xDT y yDT con llaves casteadas.
}
\description{
Se encarga de realizar el casteo entre dos data.tables de las
variables contenidas en _keys_.
}
