% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean.by.keys}
\alias{clean.by.keys}
\title{Limpia data.table de duplicados}
\usage{
clean.by.keys(dt, keys)
}
\arguments{
\item{dt}{data.table a limpiar}

\item{keys}{llaves que conforman cada grupo unico}
}
\value{
DT sin duplicados.
}
\description{
Dado un merge con elementos de llave duplicados, se encarga de
mantener solo la ultima coincidencia de los casos permutados por repeticion.
}
\note{
Se puede agregar un nuevo parametro 'keep', para determinar que elemento
conservar.
}
