% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codsetor2blockid}
\alias{codsetor2blockid}
\title{Obtener BlockId desde COD_SETOR}
\usage{
codsetor2blockid(codsetor)
}
\arguments{
\item{codsetor}{Caracter correspondiente a COD_SETOR}
}
\value{
Caracter correspondiente a BLOCKID
}
\description{
Obtiene el BLOCKID desde COD_SETOR.
}
