% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{codsetor2geo}
\alias{codsetor2geo}
\title{Define Columnas Geograficas a partir de Codigo de Sector}
\usage{
codsetor2geo(codsetor)
}
\arguments{
\item{codsetor}{Caracter correspondiente a COD_SETOR, debe contener 15 digitos.}
}
\value{
Lista con las variables ESTADO, MUNIC, COD_DIG, DISTRITO, SUBDIST, SETOR.
}
\description{
Se encarga de obtener variables geograficas de acuerdo a la
variable COD_SETOR.
}
