% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_duplicate_columns}
\alias{eval_duplicate_columns}
\title{Evalua Columnas Duplicadas en un data.table}
\usage{
eval_duplicate_columns(dt, suffixes = c(".x", ".y"))
}
\arguments{
\item{dt}{Data.table a evaluar sus columnas.}

\item{suffixes}{vector de solo dos elementos, donde se especifica como se llaman
las columnas duplicadas.}
}
\value{
Data.table con columnas evaluadas con terminacion '.equal'
}
\description{
Evalua columnas duplicadas en un data.table, producto de un merge, para
determinar si son iguales de acuerdo a su contenido y su posicion..
Busca columnas duplicadas de acuerdo a los nombres que terminen con '.x' o '.y'.
}
