% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_equal_columns}
\alias{eval_equal_columns}
\title{Evalua Columnas .equal}
\usage{
eval_equal_columns(DT)
}
\arguments{
\item{DT}{data.table resultado de la funcion _eval_duplicate_columns_,
con columnas .equal.}
}
\value{
Lista de subsets de DT, donde cada indice es una de las columnas evaluadas,
retornando solo resultados no iguales.
}
\description{
Realiza una verificacion logica entre columnas con terminacion .equal
}
