% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo2blockid}
\alias{geo2blockid}
\title{Obtiene BlockId por medio de variables geograficas.}
\usage{
geo2blockid(estado, munic, distrito, subdist, setor)
}
\arguments{
\item{estado}{id del estado}

\item{munic}{id del municipio}

\item{distrito}{id del distrito}

\item{subdist}{id del subdistrito}

\item{setor}{id del sector}
}
\value{
BLOCKID
}
\description{
Genera BLOCKID de acuerdo a las variables geograficas estado,
municipio, distrito, subdistrito y sector.
}
