% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo2codsetor}
\alias{geo2codsetor}
\title{Obtiene COD_SETOR desde Geografia}
\usage{
geo2codsetor(ESTADO, MUNIC, COD_DIG, DISTRITO, SUBDIST, SETOR)
}
\arguments{
\item{estado}{id del estado}

\item{munic}{id del municipio}

\item{distrito}{id del distrito}

\item{subdist}{id del subdistrito}

\item{setor}{id del sector}
}
\value{
BLOCKID
}
\description{
Genera BLOCKID de acuerdo a las variables geograficas estado,
municipio, distrito, subdistrito y sector.
}
