% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gps2lat_lon}
\alias{gps2lat_lon}
\title{Obtiene Latitud y Longitud de variables GPS}
\usage{
gps2lat_lon(x, y)
}
\arguments{
\item{x}{vector caracter representando GPS1}

\item{y}{vector caracter representando GPS2}
}
\value{
lista con valores Latitud y Longitud
}
\description{
Obtiene la Latitud y Longitud de la variable GPS y GPS2.
}
