% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_allowed}
\alias{is_allowed}
\title{¿Permitido?}
\usage{
is_allowed(factor, var)
}
\arguments{
\item{factor}{vector numerico representando el factor}

\item{var}{vector numerico representando la variacion porcentual (en 100)}
}
\value{
valor logico de la evaluacion
}
\description{
Funcion predicado que sirve para evaluar si una combinacion de
factor y variacion esta en el rango permitido.
}
