% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iterative.merge}
\alias{iterative.merge}
\title{Merge iterativo}
\usage{
iterative.merge(
  refDT,
  list.DT,
  list.keys,
  ...,
  keep.y = T,
  solve_conflict = c("erase", "substitute")
)
}
\arguments{
\item{refDT}{data.table de referencia.}

\item{list.DT}{Lista de data.table's para aplicar en refDT.}

\item{list.keys}{Lista de vectores de caracteres correspondiente a las llaves.}

\item{...}{Calquier parametro extra a aplicar en 'merge'.}

\item{keep.y}{Se utiliza para mantener prioridad a la nueva informacion entrante.
Esto debido a que se usa como parametro dentro del flujo de la funcion, en las
funciones 'prepare_dts', 'merge_substitute', 'rid_duplicate_columns'.}

\item{solve_conflict}{decide si los conflictos en cada merge se resuelven
o bien manteniendo la ultima columna ingresada o substituyendo en donde el valor
sea nuevo.}
}
\value{
data.table resultado de todos los merge.
}
\description{
Realiza merges consecutivos teniendo como primera referencia refDT
por cada data.table contenido en list.DT, tomando como llaves
cada indice de list.keys. A refDT se le aplica un merge del primer elemento de
listDT, usando como llave el primer elemento de list.keys y asi consecutivamente.
}
