% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_pkgs}
\alias{load_pkgs}
\title{Carga o Instala Paqueterias}
\usage{
load_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Vector de caracteres con el nombre de paqueterias a cargar.}
}
\value{
No hay retorno, solo hay una descarga de elementos
}
\description{
Se encarga de hacer carga de todas las paqueterias contenidas
en _pkgs_. Se encarga de instalar cualquier paqueteria no instalada.
}
\examples{
\dontrun{ load_pkgs(c("data.table", "purrr", "stringr", "futile.logger",
"foreach", "tidyr", "haven", "openxlsx",
"future", "furrr", "glue", "bit64", "feather"))
}
}
