% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_msg}
\alias{log_msg}
\title{Envia mensaje al sistema de loggeo}
\usage{
log_msg(msg, fun = flog.info)
}
\arguments{
\item{msg}{Mensjaje a imprimir en logs.}

\item{fun}{Funcion del nivel de loggeo. Por default se usa __flog.info__, pero se
puede usar tambien __flog.warn__, __flog.debug__, __flog.error__, __flog.fatal__, etc.}
}
\value{
No hay retorno, solo se concatena el mensaje en el archivo de loggeo y
en la consola
}
\description{
Una vez que el sistema de loggeo se ha inicializado, por medio de esta
funcion se puede imprimir mensajes de loggeo. Los mensajes apareceran tanto en el
archivo como en la consola.
}
