% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.list}
\alias{merge.list}
\title{Aplica un merge a una lista}
\usage{
\method{merge}{list}(list.dt, key, ...)
}
\arguments{
\item{list.dt}{lista con data.tables a unificar}

\item{key}{caracter con el nombre de la llave que comparten todos los data.tables}

\item{...}{parametros extras a ingresar en \code{Reduce}}
}
\description{
Version simplificada de **iterative_merge** pero que usa solo una lista
de data.tables para iterar sobre si misma. Esta funcion, a diferencia de su original
pretende eliminar generaciones de parametros grandes. Conteniendo a su vez la misma
funcion de limpieza de columnas duplicadas y una sola llave compartida.
}
