% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_columns}
\alias{merge_columns}
\title{Unifica columnas}
\usage{
merge_columns(dt, keep.y = TRUE, suffixes = c(".x", ".y"))
}
\arguments{
\item{dt}{data.table ca aplicar funcion}

\item{keep.y}{Si se mantiene prioridad en columna 2 (TRUE) o en columna 1 (FALSE)}

\item{suffixes}{Vector de 2 caracteres que representan los sufijos de las columnas
duplicadas. Por default es '.x' y '.y' ya que es el resultado por defecto de merge.}
}
\value{
data.table sin columnas duplicadas
}
\description{
Unifica los valores de columnas duplicadas resultado de un merge.
}
