% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_substitute}
\alias{merge_substitute}
\title{Merge subtitutivo}
\usage{
merge_substitute(DT1, DT2, unique.key, keep.y = T)
}
\arguments{
\item{DT1}{data.table de referencia}

\item{DT2}{data.table catalogo, o de quien se tomaran los nuevos valores}

\item{unique.key}{Vector de caracteres usados como llaves.}

\item{keep.y}{Si se mantendra prioridad a DT2 para substituir los valores de DT1}
}
\value{
data.table con valores de llave en comun subtituidos
}
\description{
Realiza una subtitucion de valores dado un merge. Se utiliza para
evitar duplicidad de columnas con nombres iguales y si solo se quiere modificar
variables en comun dado otro data.table.
}
