% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{nielsenUtils}
\alias{nielsenUtils}
\alias{nielsenUtils-package}
\title{nielsenUtils: Paqueteria de funciones utilitarias comunes
en GDDC Brasil Nielsen Mexico}
\description{
Esa paqueteria contiene un conjunto de funciones, variables y procesos
que son comunes en distintos procesos y tareas de GDDC Brasil Nielsen,
por lo que se engloban en esta paqueteria para poder mantenerse de
forma facil y compartirse con el equipo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielQuinmat/nielsenUtils}
  \item Report bugs at \url{https://github.com/GabrielQuinmat/nielsenUtils/issues}
}

}
\author{
Gabriel Quintanar \email{yahirgabriel.quintanarmartinez@nielsen.com}
}
