% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_char_columns}
\alias{parse_char_columns}
\title{Formatea vectores de caracteres}
\usage{
parse_char_columns(dt, format = TRUE)
}
\arguments{
\item{dt}{data.table a aplicar funcion}

\item{format}{Si se convierte en mayuscula todas las columnas caracter (TRUE) o no
(FALSE).}
}
\value{
data.table formateado
}
\description{
Se encarga de substituir caracteres especiales englobados en la
variable global special_chrs solo a variables caracter en _dt_.
Tambien puede cambiar el texto a mayusculas.
}
