% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_product}
\alias{pivot_product}
\title{Pivotear un producto Nielsen}
\usage{
pivot_product(
  var.mm,
  y.vars,
  value,
  dt,
  func,
  ...,
  keeps = "manejante|no_manejante|porcentaje|total"
)
}
\arguments{
\item{var.mm}{Variable correspondiente al producto a pivotear. Puede ser multinivel
con alguna variable categorica.}

\item{y.vars}{Variables correspondiente al nivel de agregacion categorica.}

\item{value}{Variable que se usara para agregarse.}

\item{dt}{Data.table del cual se recoge la informacion.}

\item{func}{Funcion de agregacion.}

\item{...}{Parametros extras que pueden pasarse a la funcion de agregacion.}

\item{keeps}{Variables a mantener del resultado. Las variables resultantes son:
\itemsize{
\item .manejante: Cantidad de manejantes.
\item .no_manejante: Cantidad de no manejantes.
\item .porcentaje: Porcentaje de manejantes.
\item .total: Cantidad total.
}
Por default se extraen las 4 variables.}
}
\value{
data.table con informacion pivoteada del producto.
}
\description{
Funcion pivoteadora de un solo producto. Ejecuta un dcast internamente
de una columna logica correspondiente a un producto. Pivotea los valores para poder
determinar las cantidades manejantes, porcentajes y totales.
}
