% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_dts}
\alias{prepare_dts}
\title{Prepara data.tables para un merge}
\usage{
prepare_dts(xDT, yDT, keys, keep.y = T)
}
\arguments{
\item{keys}{Vector de caracteres correspondiente a las llaves a utilizar en merge.}

\item{ref}{DT de referencia.}

\item{dt}{DT a unir.}
}
\value{
Objeto lista con datatables ref y dt y las llaves.
}
\description{
Prepara dataframes para la funcion __recursive_merge__. Este
procesamiento consiste solo en revisar llaves para el merge para formatear
las columnas al mismo tipo; tambien se encarga de quitar columnas del DT
de referencia y conservar otras columnas del DT a unir.
}
