% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dup.verification}
\alias{print.dup.verification}
\title{Imprime una verificacion de duplicacion}
\usage{
\method{print}{dup.verification}(dt, key)
}
\arguments{
\item{dt}{data.table a verificar}

\item{key}{llave a usar para verificar}
}
\value{
No hay retorno, solo mensajes a consola
}
\description{
Imprime informacion de tuplas repetidas por generacion de merge en el
sistema de loggeo.
}
