% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{produce_formula}
\alias{produce_formula}
\title{Genera un objeto Formula}
\usage{
produce_formula(lhs, rhs)
}
\arguments{
\item{lhs}{Left hand side: Vector de caracteres correspondientes al lado izquierdo
de la formula.}

\item{rhs}{Right hand side:, Vector de caracteres correspondientes al lado derecho
de la formula.}
}
\value{
Formula sin evaluar.
}
\description{
Genera un objeto formula desde un string que tenga el sintaxis de una formula.
}
