% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{product_pivotizer}
\alias{product_pivotizer}
\title{Pivotea productos}
\usage{
product_pivotizer(
  products,
  keys,
  value.var,
  func,
  ...,
  keeps = "manejante|no_manejante|porcentaje|total",
  dt
)
}
\arguments{
\item{products}{Vector de caracteres correspondientes a los productos a pivotear.}

\item{keys}{Niveles de agregacion para pivotear los productos. Posteriormente se
convierten en las llaves a utilizar en la union final de los pivotes.}

\item{value.var}{Columna de valores a agregar.}

\item{func}{Funcion de agregacion a utilizar en el pivoteo.}

\item{...}{Parametros extras a utilizar en la funcion de agregacion __func__.}

\item{keeps}{Vector de caracteres correspondiente a variables a mantener de cada
pivoteo de los productos. Para mas detalles, refierase a funcion __pivot_product__,}

\item{dt}{Data.table de referencia.}
}
\value{
Tabla completa con pivoteo de productos.
}
\description{
Funcion de alto nivel para poder pivotear varios productos a la vez.
Utiliza iterativamente la funcion __pivot_product__ para generar pivotes de productos.
Posteriormente los une por medio de un merge para la tabla final.
}
