% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.dt}
\alias{read.dt}
\title{Lee cualquier archivo de datos como data.table}
\usage{
read.dt(path, col_class = "auto", .verbose = F, ...)
}
\arguments{
\item{path}{Ruta del archivo a leer.}

\item{col_class}{Se puede delegar la transformacion de columnas a la paqueteria
correspondiente con 'auto', o se pueden convertir todas a String con 'chr' y
despues castear.}

\item{.verbose}{valor logico que define si se mandaran mensajes logs acerca
de la lectura y transformacion de datos.}

\item{...}{Cualquier parametro extra que se pueda usar en 'read_sas', 'read.xlsx',
'fread', 'readRDS' y 'read_feather'.}
}
\value{
data.table resultante.
}
\description{
Se encarga de leer un archivo compatible (texto, excel o SAS),
al leer la terminacion del archivo. A su vez, convierte a mayusculas todos
los headers y elimina caracteres especiales.
}
