% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rid_duplicate_columns}
\alias{rid_duplicate_columns}
\title{Elimina columnas duplicadas}
\usage{
rid_duplicate_columns(dt, keep.y)
}
\arguments{
\item{dt}{data.table al que se le acaba de aplicar un merge y resulto con
columnas duplicadas}

\item{keep.y}{valor logico para decidir si se conserva el valor nuevo \code{TRUE}
o el valor viejo \code{FALSE}}
}
\value{
data.table sin columnas duplicadas por merge
}
\description{
Unifica columnas duplicadas debido al merge, cuya extension se vuelve
'.x' y '.y'. Mantiene prioridad a la columna con terminacion '.y' ya que es quien
contiene informacion nueva.
}
\note{
Se puede generar un nuevo parametro 'priority' para determinar quien tiene la
relevancia.
}
