% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.percent}
\alias{set.percent}
\title{Calcula porcentajes logicos}
\usage{
set.percent(true, false)
}
\arguments{
\item{true}{Vector numerico correspondiente a verdaderos.}

\item{false}{Vector numerico correspondiente a falsos.}
}
\value{
Lista con resultados de cantidades verdaderas, falsas, porcentaje de
verdaderos y total.
}
\description{
Computa el porcentaje de acuerdo a dos variables logicas.
}
