% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_logger}
\alias{set_logger}
\title{Inicializa un sistema de loggeo}
\usage{
set_logger(log.file = "logger.log")
}
\arguments{
\item{log.file}{Define el nombre del archivo .log, por defecto es logger.log}
}
\value{
No hay retorno, solo inicaliza el sistema
}
\description{
Inicializa un doble sistema de loggeo. Uno para un archivo, con nombre
logger.log, y otro para consola. El sistema de loggeo por archivo se inicializa
en la carpeta del working directory.
}
