% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{special_chrs}
\alias{special_chrs}
\title{#' @description Se utiliza como diccionario para eliminar caracteres especiales
special_chrs <- c('Š'='S', 'š'='s', 'Ž'='Z', 'ž'='z', 'À'='A', 'a'='A', 'Â'='A',
                 'Ã'='A', 'Ä'='A', 'Å'='A', 'Æ'='A', 'Ç'='C', 'È'='E', 'e'='E',
                 'Ê'='E', 'Ë'='E', 'Ì'='I', 'i'='I', 'Î'='I', 'Ï'='I', 'Ñ'='N',
                 'Ò'='O', 'o'='O', 'Ô'='O', 'Õ'='O', 'Ö'='O', 'Ø'='O', 'Ù'='U',
                 'u'='U', 'Û'='U', 'Ü'='U', 'Ý'='Y', 'Þ'='B', 'ß'='Ss', 'à'='a',
                 'a'='a', 'â'='a', 'ã'='a', 'ä'='a', 'å'='a', 'æ'='a', 'ç'='c',
                 'è'='e', 'e'='e', 'ê'='e', 'ë'='e', 'ì'='i', 'i'='i', 'î'='i',
                 'ï'='i', 'ð'='o', 'ñ'='n', 'ò'='o', 'o'='o', 'ô'='o', 'õ'='o',
                 'ö'='o', 'ø'='o', 'ù'='u', 'u'='u', 'û'='u', 'ý'='y', 'ý'='y',
                 'þ'='b', 'ÿ'='y' )
Vector de caracteres especiales}
\format{
Vector con caracteres especiales:
\describe{
  \item{special_chrs}{Valores de alfabeto especial}
}
}
\source{
Creado
}
\usage{
special_chrs
}
\description{
Vector con nombre, que conserva todos los caracteres especiales existentes
en codificacion Latin-1. En el nombre contiene todos sus valores estandarizados.
}
\keyword{datasets}
