% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn.num}
\alias{warn.num}
\title{Casteo a numerico con advertemcia}
\usage{
warn.num(x)
}
\arguments{
\item{x}{vector a castear a numerico}
}
\value{
vector numerico casteado
}
\description{
Parsea texto a numero y detiene si hay texto no parseable
}
\details{
Si el vector contiene elementos no parseables, envia una advertencia
con el valor de dichos elementos.
}
