% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{special_chrs}
\alias{special_chrs}
\title{Vector de caracteres especiales}
\format{
Vector con caracteres especiales:
\describe{
  \item{special_chrs}{Valores de alfabeto especial}
}
}
\source{
Creado
}
\usage{
special_chrs
}
\description{
Vector con nombre, que conserva todos los caracteres especiales existentes
en codificacion Latin-1. En el nombre contiene todos sus valores estandarizados.
}
\keyword{datasets}
